Split Cardano module into smaller modules
I will be adding many things to interact with the wallet, and I don't
want all of it to be in a single big module, so let's make
Adatipd.Cardano a group for the parts.
